requirements.txt
customtkinter>=5.2.0
cryptography>=41.0.0
pillow>=10.0.0
pystray>=0.19.0
